Use ssize_t for array size
`-1` indicates size is unspecified (e.g. `int array[]`).

+++ debug/dump_type.c
@@ -46,7 +46,7 @@ void dump_type(FILE *fp, const Type *type) {
     break;
   case TY_ARRAY:
     fprintf(fp, "Array<");
-    if (type->pa.length != (size_t)-1)
+    if (type->pa.length != -1)
       fprintf(fp, "%zu, ", type->pa.length);
     else
       fprintf(fp, "?, ");
